Refactor Lorre to make it more generic (move all tezos classes to tezos package)
