+++ conseil-lorre/src/main/scala/tech/cryptonomic/conseil/indexer/tezos/TezosErrors.scala
@@ -1,7 +1,7 @@
-package tech.cryptonomic.conseil.common.tezos
+package tech.cryptonomic.conseil.indexer.tezos
 
 /** Defines high-level errors encountered during processing */
-trait TezosErrors {
+private[tezos] object TezosErrors {
 
   /** Something went wrong during handling of Blocks or related sub-data */
   case class BlocksProcessingFailed(message: String, cause: Throwable) extends java.lang.RuntimeException
