+++ conseil-lorre/src/test/scala/tech/cryptonomic/conseil/indexer/tezos/TezosNodeOperatorTest.scala
@@ -7,7 +7,6 @@ import org.scalatest.{FlatSpec, Matchers}
 import tech.cryptonomic.conseil.common.config.BatchFetchConfiguration
 import tech.cryptonomic.conseil.common.tezos.TezosTypes
 import tech.cryptonomic.conseil.common.tezos.TezosTypes.BlockHash
-import tech.cryptonomic.conseil.indexer.LorreOperations
 
 import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
@@ -32,7 +31,7 @@ class TezosNodeOperatorTest
         .when("zeronet", "blocks/BLockGenesisGenesisGenesisGenesisGenesisb83baZgbyZe~")
         .returns(blockResponse)
 
-      val apiOps: LorreOperations = new LorreOperations
+      val apiOps: TezosNodeOperations = new TezosNodeOperations
       val nodeOp: TezosNodeOperator = new TezosNodeOperator(tezosRPCInterface, "zeronet", config, apiOps)
 
       //when
@@ -84,7 +83,7 @@ class TezosNodeOperatorTest
             )
         )
 
-      val apiOps: LorreOperations = new LorreOperations
+      val apiOps: TezosNodeOperations = new TezosNodeOperations
       val nodeOp: TezosNodeOperator = new TezosNodeOperator(tezosRPCInterface, "zeronet", config, apiOps)
 
       //when
@@ -146,7 +145,7 @@ class TezosNodeOperatorTest
             )
         )
 
-      val apiOps: LorreOperations = new LorreOperations
+      val apiOps: TezosNodeOperations = new TezosNodeOperations
       val nodeOp: TezosNodeOperator = new TezosNodeOperator(tezosRPCInterface, "zeronet", config, apiOps)
 
       //when
@@ -207,7 +206,7 @@ class TezosNodeOperatorTest
             )
         )
 
-      val apiOps: LorreOperations = new LorreOperations
+      val apiOps: TezosNodeOperations = new TezosNodeOperations
       val nodeOp: TezosNodeOperator = new TezosNodeOperator(tezosRPCInterface, "zeronet", config, apiOps)
 
       //when
