+++ docker/entrypoint.sh
@@ -11,6 +11,7 @@ XTZSCHEME=${XTZ_Scheme:=http}
 XTZHOST=${XTZ_Host:=node}
 XTZPREFIX=${XTZ_Prefix:=}
 XTZPORT=${XTZ_Port:=8732}
+XTZPLATFORM=${XTZ_Platform:=tezos}
 XTZNET=${XTZ_Network:=mainnet}
 APIPORT=${API_PORT:=80}
 APIKEY=${API_KEY:=conseil}
@@ -35,6 +36,7 @@ if [ $CONFIG = "none" ]; then
     sed -i "s/{{XTZHOST}}/$XTZHOST/g" conseil.conf
     sed -i "s/{{XTZPREFIX}}/$XTZPREFIX/g" conseil.conf
     sed -i "s/{{XTZPORT}}/$XTZPORT/g" conseil.conf
+    sed -i "s/{{XTZPLATFORM}}/$XTZPLATFORM/g" conseil.conf
     sed -i "s/{{XTZNET}}/$XTZNET/g" conseil.conf
     sed -i "s/{{APIPORT}}/$APIPORT/g" conseil.conf
     sed -i "s/{{APIKEY}}/$APIKEY/g" conseil.conf
@@ -48,5 +50,5 @@ if [ $1 = "conseil-api" ]; then
 fi
 
 if [ $1 = "conseil-lorre" ]; then
-  java -Xmx$JVM_XMX_LORRE -Dconfig.file=$CONFIG -cp /root/conseil-lorre.jar tech.cryptonomic.conseil.indexer.Lorre $XTZNET
+  java -Xmx$JVM_XMX_LORRE -Dconfig.file=$CONFIG -cp /root/conseil-lorre.jar tech.cryptonomic.conseil.indexer.Lorre $XTZPLATFORM $XTZNET
 fi
