+++ docker/template.conf
@@ -1,7 +1,7 @@
 include required(classpath("application.conf"))
 
 platforms: {
-  tezos: {
+  {{XTZPLATFORM}}: {
     {{XTZNET}}: {
       node: {
         protocol: "{{XTZSCHEME}}"
